src/aws_control/requirements.txt
absl-py==1.3.0
actionlib==1.14.0
addict==2.4.0
aiobotocore==1.2.0
aiodns==2.0.0
aiofiles==0.4.0
aiohttp==3.7.4.post0
aioitertools==0.7.1
aiosignal==1.3.1
angles==1.9.13
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asteval==0.9.21
asttokens==2.2.1
async-timeout==3.0.1
attrs==20.1.0
autopep8==2.0.1
backcall==0.2.0
beautifulsoup4==4.11.1
bezier==2020.5.19
bleach==5.0.1
bondpy==1.8.6
boto3==1.17.1
botocore==1.20.49
brotlipy==0.7.0
cachetools==5.2.0
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==0.5.2
cchardet==2.1.6
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.4
click==7.1.2
comm==0.1.2
ConfigArgParse==1.5.3
construct==2.10.61
contourpy==1.0.7
controller-manager==0.19.6
controller-manager-msgs==0.19.6
coverage==6.1.2
cryptography @ file:///croot/cryptography_1665612644927/work
cupy-cuda11x==11.4.0
cv-bridge==1.16.2
cycler==0.10.0
Cython==0.29.33
dash==2.8.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
descartes==1.1.0
dgl==0.9.1
diagnostic-analysis==1.11.0
diagnostic-updater==1.11.0
distlib==0.3.6
distro==1.8.0
docker-pycreds==0.4.0
docutils==0.19
dubins==1.0.1
dynamic-reconfigure==1.7.3
empy==3.3.4
entrypoints==0.4
executing==1.2.0
fastapi==0.68.1
fastjsonschema==2.16.3
fastrlock==0.8.1
filelock==3.8.0
fire==0.4.0
Flask==2.2.3
fonttools==4.39.0
frozenlist==1.3.3
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
gitdb==4.0.10
GitPython==3.1.29
google-auth==2.14.1
google-auth-oauthlib==0.4.6
grpcio==1.50.0
gunicorn==20.1.0
h11==0.14.0
httptools==0.1.1
idna==3.4
idna-ssl==1.1.0
image-geometry==1.16.2
imageio==2.22.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
interactive-markers==1.12.0
ipykernel==6.21.3
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.2.0
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.3.2
laser_geometry==1.6.7
llvmlite==0.37.0
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.4.3
matplotlib-inline==0.1.6
message-filters==1.16.0
mistune==2.0.4
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
msgpack==1.0.4
multidict==4.7.6
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.1
networkx @ file:///home/conda/feedstock_root/build_artifacts/networkx_1667391556261/work
notebook==6.5.2
notebook_shim==0.2.2
numba==0.51.2
numpy==1.23.5
numpy-quaternion==2020.9.5.14.42.2
nuscenes-devkit==1.1.9
nvsmi==0.4.2
oauthlib==3.1.1
open3d==0.17.0
opencv-python==4.6.0.66
osqp==0.6.2.post5
packaging==21.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
Pint==0.17
platformdirs==2.6.2
plotly==5.13.1
plyfile==0.7.4
positional-encodings==6.0.1
prometheus-client==0.15.0
promise==2.3
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.7.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==3.1.1
pycocotools==2.0.6
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.9.2
Pygments==2.14.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==2.4.7
pyperclip==1.8.0
pyquaternion==0.9.9
pyrsistent==0.19.2
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.8.2
python-multipart==0.0.5
python-qt-binding==0.4.4
pytz==2022.7.1
PyYAML==5.4.1
pyzmq==25.0.0
qdldl==0.1.5.post0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
qtconsole==5.4.0
QtPy==2.3.0
ray==2.1.0
requests==2.26.0
requests-oauthlib==1.3.0
resource_retriever==1.12.7
rosapi==0.11.16
rosbag==1.16.0
rosboost-cfg==1.15.8
rosbridge-library==0.11.16
rosbridge-server==0.11.16
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.16.0
roslaunch==1.16.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.16.0
rosmake==1.15.8
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospkg==1.5.0
rospy==1.16.0
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rqt-image-view==0.4.17
rqt-moveit==0.5.10
rqt-reconfigure==0.5.5
rqt-robot-monitor==0.5.14
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.11
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.9
rqt_logger_level==0.4.11
rqt_msg==0.4.10
rqt_plot==0.4.13
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_runtime_monitor==0.5.9
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_tf_tree==0.6.3
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rsa==4.9
rviz==1.14.20
s3transfer==0.3.7
scikit-learn==1.2.1
scipy==1.7.1
selenium==3.141.0
Send2Trash==1.8.0
sensor-msgs==1.13.1
sentry-sdk==1.9.8
setproctitle==1.3.2
Shapely==1.8.5.post1
shortuuid==1.0.11
six==1.16.0
smach==2.5.1
smach-ros==2.5.1
smclib==1.8.6
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
starlette==0.14.2
tenacity==8.2.2
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.1.1
terminado==0.17.0
tf==1.13.2
tf2-geometry-msgs==0.7.6
tf2-py==0.7.6
tf2-ros==0.7.6
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.1
tomli==2.0.1
topic-tools==1.16.0
torch==1.13.1+cu117
torch-tb-profiler==0.4.0
torchvision==0.14.0+cu117
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing-extensions==3.7.4.3
urllib3==1.26.12
uvicorn==0.12.1
virtualenv==20.17.0
wandb==0.13.5
watchdog==0.10.4
watchgod==0.7
wcwidth==0.2.6
webbot==0.34
webencodings==0.5.1
websocket-client==1.4.2
websockets==9.1
Werkzeug==2.2.3
widgetsnbextension==4.0.5
wincertstore==0.2
wrapt==1.10.10
xacro==1.14.16
yarl==1.5.1
zipp==3.10.0
